twin-orchestrator/requirements.txt
# Twin Orchestrator Dependencies

# HTTP client for TuringCore API
httpx>=0.25.0

# YAML configuration parsing
pyyaml>=6.0.1

# Data validation and serialization
pydantic>=2.5.0

# CLI framework (for future CLI commands)
click>=8.1.7

# Progress bars and logging
rich>=13.7.0

# Date/time utilities
python-dateutil>=2.8.2

# Faker for synthetic data generation (v0.2+)
# faker>=20.1.0  # Uncomment when implementing realistic data generators

# TuringCore API client (when published)
# turingcore-client>=1.0.0  # Uncomment when OpenAPI client SDK is available
